pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "luanny"
version = "0.1.0"
description = "Coletor de dados de acessibilidade do Instagram para pesquisa acadêmica"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Projeto Luanny" },
]
dependencies = [
    "playwright>=1.40",
    "pydantic>=2.0",
    "typer[all]>=0.9",
    "pandas>=2.0",
    "python-dotenv>=1.0",
    "structlog>=23.0",
    "tenacity>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1,<0.7",
]
build = [
    "pyinstaller>=6.0",
]

[project.scripts]
luanny = "luanny.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/luanny"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that hit live Instagram (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
